* Replace imake with configure * Many portability kludges replaced by <sysdep.h> * Added John Hawkinson's Solaris streams logging driver.
Here are the notes from jhawk on his streams log driver:

These patches add support for the Solaris-style STREAMS-based /dev/log
interface to the Zephyr syslogd.

The relevent changes are #ifdef'd on STREAMS_LOG_DRIVER, which in
turn is defined if SOLARIS is defined.

Use of streams requires that poll() be used rather than select(),
so there are some crappy #ifdef's littered throughout there.
It seems nonobvious how to divide them up in a more manageable
fashion, so it's left as is.

This code assumes that NLOGARGS will be three, and #error's if
that's not the case. This is due to a bogosity in the log(7)
specification (where the /dev/log interface is documented) which
specifies that it returns NLOGARGS words which much be *printf()'d,
which is not possible to do in a portable fashion without either:

	1) a portable interface to the stdarg back end
	2) preprocessor and sed magic

Since I chose to avoid the latter, and the former does not
exist, this restriction is in place.

Some other notes:

*	/dev/klog is not supported in this model (doesn't exist).

*	There are two other types of logging, I_ERRLOG and I_TRCLOG. It's
	not clear to me exactly what they're good for, but they provide
	far more information than is normally availabel through syslogs
	(largely stats and information from kernel modules, I believe).

*	unlink()-ing /dev/log (as syslogd normally does) is a bad idea
	under Solaris, so don't do it :-). If you do this by accident, you
	can remake it by running /usr/sbin/devlinks.