Underpinnings for realm.list enhancements
Replace the per-realm array of servers with a per-realm array of
struct _ZRealm_server, so that we can have additional information
associated with each server.

Additionally, introduce the concept that not every server in a realm's
list is necessarily a suitable place to send notices.  This means that
when selecting a server, we may need to skip ineligible entries.