server: avoid blocking in tkt_lookup()
tkt_lookup() is supposed to quickly obtain a ticket for a foreign realm
if we already have a usable one, and quickly fail otherwise.  Sending a
request to a KDC and waiting for a response, as krb5_get_credentials()
may do, defeats the purpose of tkt_retrieve() retrying failed requests
in the background.  So, use krb5_cc_retrieve_cred() instead.

Extracted from Andrew zephyr/063