More i18n infrastructure
lib/charset.c
h/zephyr/zephyr.h
Refactor ZGetCharset into ZGetCharset and a function that does
all the same defaulting but returns a string.

ZTransliterate is a zephyr-metaphor wrapper for iconv.

lib/zephyr-tests.py
Wrappers for the functions in charset.c (but not tests)

lib/zephyr_tests.txt
Tests for the functions in charset.c