Strip the authentication in realm_sendit
realm_sendit is responsible for sending notices that do not have useful
realm authentication, either because they are not authentic, or because
of kerberos problems acquiring a ticket for the foreign zephyr realm.  In
either case, any authentication in the notice will not be usable to the
foreign server, and ought to be stripped out.