Introduce usable flag on other-realm servers
Introduce a new per-realm-server 'usable' flag, which indicates the entry has
been fully initialized and can be used.  Routines which select a server or
attempt to find one based on its address should ignore servers on which the
usable flag is not set.

This will allow the introduction of features which require recording servers
which are not yet usable, such as asynchronous server name resolution.