Use interned strings for realm and server names
Use interned strings instead of strdup() and fixed-size arrays for names
of other realms and their servers.  This gives immediate improvement in
the form of doing fewer string compares when loading the realm.list,
plus the obvious benefit of eliminating some fixed char arrays.  It also
paves the way for efficiently identifying existing realms and servers to
be updated when reloading the realm.list.