Support calling realm_init() more than once
This makes realm_init() augment the existing other-realm array instead
of replacing it wholesale, which makes it safe to call more than once.
During the first call in which the realm.list file exists and contains at
least one realm, the otherrealms array will be initialized with entries
for all configured realms.  During subsequent calls, any new realms will
be added, growing the array as necessary.  For now, entries for existing
realms are not updated in any way.