Abstract realm server usability test
Whether a particular server is usable currently depends only on whether we
have successfully obtained its IP address.  However, eventually it will also
depend on additional factors such as whether the server has been deleted from
the realm.list.  To ease that transition, replace the 'usable' flag with a
'got_addr' flag (which is set when the address lookup succeeds), and add a
new function to test whether a server is usable.