configure: Don't demand readline, curses for libss
libss may not have direct dependencies on readline and/or curses, and
those libraries may in fact not even exist.  If there is a dependency,
a libss shared-library won't need us to link against it in any event.

While old versions of libss were normally built only as archive libraries,
I have been unable to find a version of that vintage which depends on
readline or curses.  Newer versions ship as shared libraries, and at least
one widely-ditributed implementation still has no such dependency, but
does appear to be able to load libreadline dynamically if it is present
at runtime.

KTH's libsl, which is an enhanced but drop-in-compatible libss replacement
distributed with Heimdal, does depend on their editline library.  But,
while libeditline is a functional replacement for readline, its API is
completely different.  And again, this is generally a shared-library
dependency of which we need not be aware.