server: Send SERVACK in response to GIMMESUBS
When processing a request to retrieve subscriptions, we actually do need
to send a SERVACK in response to the incoming notice before sending the
subscription list (directly) to the client that requested it.  The line
that did this was inadvertently removed when the OLD_COMPAT support was
removed in commit 9b709859db5310444052d13ed8ebccec6ead1669.

This fixes #91