Treat obviously non-zephyr packets as bad
When Z_ReadWait receives a packet which doesn't start with a zephyr
version header, it considers the packet to be "obviously non-zephyr".
Such packets are discarded and, previously, caused Z_ReadWait to return
ZERR_NONE.  Unfortunately, this can cause things to block for up to 60s
when a caller was expecting a non-blocking call to pick up a new packet
if there is one.

This changes Z_ReadWait to return ZERR_BADPKT in this situation,
eliminating the potential wait.

This fixes #100