Make ZRealm.srvrs be an array of pointers
With asynchronous name resolution and timers, we need to keep around
pointers to individual other-realm servers.  This, we cannot move
existing servers around in memory without causing data corruption.
But, realm_init() wants to reallocate the srvrs array for a realm when
adding servers.

Therefore, to allow ZRealm.srvrs to be reallocated without changing the
addresses of existing servers, it is converted from an array of servers
to an array of pointers to servers.