Provide a pkg-config file for zephyr
Previously, it was difficult to detect the presence of the zephyr
library in autoconf, and required custom macros.  However, the world
has since developed pkg-config, which is a simple tool for detecting
the presence of a package, its compile-time flags, and its link-time
flags, even in the presence of recursive dependencies.  This adds
"zephyr.pc" as a file generated by the build process, and installs it
into the appropriate directory, allowing the target system to use

PKG_CHECK_MODULES([ZEPHYR], [zephyr])
AC_SUBST([ZEPHYR_CFLAGS])
AC_SUBST([ZEPHYR_LIBS])

to detect all necessary information to incorporate the zephyr library.