debian: Enable seccomp
Commit 22a59d2d8b48cb2762362bd71e24293a53bca09a added basic seccomp
support to zhm. Since zhm currently runs as root and handles network
traffic, any sandboxing we can get is a good idea. Take a dependency
on libseccomp-dev and enable seccomp support for zhm. This breaks
building on kFreeBSD, but this package doesn’t currently build there
anyway, so it’s not making things any worse.